refactor(lessons): KnowledgeCheckEditor sous 300 lignes (H5)
Éclatement complet (pattern G1) : 864 → 249 lignes.
- useKnowledgeCheckEditor.js (90 l) : état du quiz, validation globale, ajout/
  suppression de questions, sauvegarde (create/update). Logique verbatim.
- QuestionsSection.vue (147 l) : liste triable (vuedraggable) + état vide + aide
  de validation ; questions en v-model.
- QuestionEditorCard.vue (197 l) : édition d'une question (en-tête, énoncé,
  explication, points) ; changement de type local.
- QuestionOptionsEditor.vue (181 l) : options + bonne réponse (radio/checkbox),
  ajout/suppression d'option (logique verbatim).
- QuizConfigForm.vue (171 l) : section configuration (scores, bascules, requis).
- Tests : composable (6) + 4 montages (15). CSS déplacé verbatim ; classes
  partagées (form-group/form-input/section-title) dupliquées par composant.
Parité : mêmes services/champs/classes ; .ghost-question placé sur la carte
(cible du ghostClass SortableJS). Aucune dette pré-existante modifiée.
